Refactors plugins
