+++ src/plugins/unpkg-path-plugin.ts
@@ -1,12 +1,6 @@
 import * as esbuild from 'esbuild-wasm';
-import axios from 'axios';
-import localForage from 'localforage';
 
-const fileCache = localForage.createInstance({
-    name: 'filecache'
-});
- 
-export const unpkgPathPlugin = (input: string) => {
+export const unpkgPathPlugin = () => {
   return {
     name: 'unpkg-path-plugin',
     setup(build: esbuild.PluginBuild) {
@@ -29,33 +23,7 @@ export const unpkgPathPlugin = (input: string) => {
             };
       });
  
-      build.onLoad({ filter: /.*/ }, async (args: esbuild.OnLoadArgs) => {
-        console.log('onLoad', args);
- 
-        if (args.path === 'index.js') {
-          return {
-            loader: 'jsx',
-            contents: input,
-          };
-        } 
 
-        const cachedResult = await fileCache.getItem<esbuild.OnLoadResult>(args.path); 
-
-        if(cachedResult) {
-            return cachedResult;
-        }
-
-        const response = await axios.get(args.path);
-        const result: esbuild.OnLoadResult = {
-            loader: 'jsx',
-            contents: response.data,
-            resolveDir: new URL('./', response.request.responseURL).pathname
-        }
-
-        await fileCache.setItem(args.path, result);
-        return result;
-        
-      });
     },
   };
 };
